services/studio/requirements.txt
flask
flask-socketio
python-dotenv
Flask-SQLAlchemy 
psycopg2-binary
flask-migrate
bcrypt
flask-jwt-extended